+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed up some typespecs to stop Elixir 1.10 complaining. More to do to make the module struct specs right.
 - Removed Timex dependency.
 - A number of bug fixes in TimeScales, plus additional type guards
+- Added basic sanitization of possible user inputs (titles, axis labels, category names from data). Sanitization approach is somewhat naive - basically any text is run through a copy of the `Plug.HTML.html_escape`.
 
 ## v0.2.0 : 2020-01-28
 - Documentation for all modules